refactor(lsp): split hover jsdoc formatting
